requirements.txt
Flask==3.0.0
psycopg2-binary==2.9.9
Werkzeug==3.0.1

Flask==3.0.0
psycopg2-binary==2.9.9
Werkzeug==3.0.1
gunicorn==21.2.0
Flask-Migrate==4.0.4